requirements.txt
Django>=6.0,<7.0
django-environ>=0.11
mysqlclient>=2.2
djangorestframework>=3.15
feedparser>=6.0
requests>=2.31
beautifulsoup4>=4.12
lxml>=5.0
anthropic>=0.40
python-dateutil>=2.9
